tx/ducktape: split streams test to its own file
